Add user authentication and authorization features
- Implement user registration and login functionality
- Create user model and routes for handling user-related requests
- Integrate JWT for secure authentication
- Update transaction handling to associate transactions with users
- Enhance API with user-specific data retrieval and management

+++ backend/modules/Transactions/deleteTransaction/Repository.ts
@@ -1,8 +1,9 @@
 import transactionModel from '../../../models/Transaction';
 
 class transactionRepository {
-	async deleteTransactionById(id: string) {
-		return transactionModel.findOneAndDelete({ id });
+	async deleteTransactionById(id	  : string, 
+								userID: string) {
+		return transactionModel.findOneAndDelete({ id, userID });
 	}
 }
